requirements.txt
google-generativeai
Pillow
autopep8
Flask
python-dotenv
gunicorn